https://online.sbis.ru/opendoc.html?guid=884e55cb-72f1-4b89-9477-4ed23133a0c9 Вынес интерфейс опций контрола, который работает с ItemActions/Controller
